MILESTONE: box64 runs a real Windows 64-bit PE end-to-end through WineGlass
A minimal 64-bit PE (Tests/mini64.c, imports GetTickCount+ExitProcess) runs to CLEAN
completion on the box64 translator via WineGlass's Win32 layer:
  GetTickCount -> 0x641EA5F1 (dispatched, returned to guest)
  guest computes 0xF1^0x2A = 0xDB = 219
  ExitProcess(219) -> 'Program exited normally', exit=0

Proves the whole pipeline on box64: x86-64 execution -> IAT/HLT thunks -> Win32
dispatch -> return-to-guest -> clean exit. The foundation for Visage is working.

box64 backend (wg_blink_box64.c) fixes that got here:
- setup_stack: real 16MB high-address guest stack (box64 pushes directly on host;
  blink's 0x7FFF0000 is unmappable on Apple Silicon)
- load_code: only set RIP when entry_rip!=0 (thunk/tramp loads passed 0, zeroing RIP)
- run: treat RIP==0 (engine's ExitProcess terminate signal) as a clean HALT, not a
  fetch-at-0 crash
- read/write_mem: sub-4GB guest accesses are not-present (read 0 / discard) — Apple
  reserves the low 4GB; only relocated high memory is real
- SIGSEGV diag handler reporting guest RIP/regs

Engine (gated WG_CPU=box64): map_initterm_tramp + s_scratch_base/cmdline/tramp/gai +
winsock gai_base widened uint32->uint64 so a rebased 64-bit image's structures aren't
truncated to unmappable low addresses.

Next: TEB/GS at a high mapped address + wg_guest_alloc >4GB for CRT init (toward Visage).

+++ Sources/Win32/wg_winsock.c
@@ -24,8 +24,8 @@
 // region here per VM load and tells us where via wg_winsock_set_gai_base()
 // (it's relocated above the image for large 64-bit PEs). Defaults to the legacy
 // low address for small/32-bit guests.
-static uint32_t s_gai_base = 0x00B00000u;
-void wg_winsock_set_gai_base(uint32_t base) { if (base) s_gai_base = base; }
+static uint64_t s_gai_base = 0x00B00000u;   // 64-bit: box64 relocates this above 4GB
+void wg_winsock_set_gai_base(uint64_t base) { if (base) s_gai_base = base; }
 
 // ── Windows socket constants ────────────────────────────────────────
 #define WSADESCRIPTION_LEN  256
